lib: Move controls to it's own folder
